requirements.txt
Django
django-rest-framework
psycopg2
factory-boy
Pillow
django-environ
